fix(web): split team workspace view state
